FastQ input accepted 	With these revisions FastQ input files will be accepted in input.list 	as long as the two paired-read files are both listed in and seperated 	by a comma

+++ scripts/Initialize/initializeLIONS.sh
@@ -126,17 +126,54 @@ do
 	rowN=$(sed -n "$nLib"p $INPUT_LIST)
 
 	# Library Name
-	libName=$(echo $rowN | cut -f1 -d' ')
+	libName=$(echo $rowN | cut -f1 -d' ')	
 
-	# Bam File Path
-	bamPath=$(echo $rowN | cut -f2 -d' ')
-	
+	# Bam File Path (or FastQ files)
+	bamPath=$(echo $rowN | cut -f2 -d' ') 
+
 	# Make library directory in project folder
 	mkdir -p $pDIR/$libName
 
-	# Link the bam file to the project folder
-	ln -fs $bamPath $pDIR/$libName/input.bam
+	# Check input files and link them to the project folder
+	if [[ $bamPath == *'.bam']]
+	then
+	# Input file is bam (standard)
+
+		if [ -s $bamPath ] #File is not empty
+		then
+		
+			# Link the bam file to the project folder
+			ln -fs $bamPath $pDIR/$libName/input.bam
+		else
+			# File is empty
+			echo ' ERROR 7A: Input Bam File is not readable / empty'
+			echo " file: $bamPath"
+			echo '  a) The input file (.bam or .fq_1 & .fq_2) isnt found'
+			echo '  b) If youre using FASTQ; the input name for the read pairs'
+			echo '    should be suffixed with _0'
+			exit 7
+		fi
+
+	else
+	# Input file should then be fastqs (comma seperated)
+		fq1=$(echo $bamPath | cut -f1 -d',' - )
+	 	fq2=$(echo $bamPath | cut -f2 -d',' - )
+
+		if [ -s $fq1 ] && [ -s $fq2 ] # Both files exist/ aren't empty
+		then
+			# Link both fastq files as input.fq_1 and input.fq_2
+			ln -fs $fq1 $pDIR/$libName/temp.1.fq.gz
+			ln -fs $fq2 $pDIR/$libName/temp.2.fq.gz
 	
+		else
+			echo ' ERROR 7B: Input File Not Accesible (Fastq)'
+			echo " files: $fq1 ; $fq2"
+			echo '  a) The non-bam input file (.fq_1 & .fq_2) isnt found'
+			echo '  b) If youre using FASTQ; make sure youre listing two'
+			echo '     files in the input.list file seperated by a comma'
+			exit 7
+		fi
+	fi	
 done
 
 # Initialize parameter folder  ******** NOT IMPLEMENTED YET *********
